+++ src/inputs.sh
@@ -2,21 +2,22 @@
 
 # Initialize variables
 help=0
-# debug=0
+debug=0
 version=0
 fail_on_not_compliant=0
 dir=
 config_json_path=
 docs_link=
+set_github_actions_outputs=0
 
 for arg in "$@"; do
   case $arg in
   -h | --help)
     help=1
     ;;
-#  -D | --debug)
-#    debug=1
-#    ;;
+  -D | --debug)
+    debug=1
+    ;;
   -v | --version)
     version=1
     ;;
@@ -31,6 +32,9 @@ for arg in "$@"; do
     ;;
   -dl=* | --docs-link=*)
     docs_link="${arg#*=}"
+    ;;
+  -gh | --github-actions)
+    set_github_actions_outputs=1
     ;;
   esac
   shift