+++ CHANGELOG.md
@@ -4,9 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [v0.9.0] - 12-11-2022
+### Added
+- Added support for analysing resources and checking naming conventions for them
+- Input for enabling optional GitHub Actions features `github_actions`
+- Input for enabling optional debug mode
+### Changed
+Nothing
+### Fixed
+Nothing
+
 ## [v0.8.0] - 04-05-2022
 ### Added
-- Comments for improving the understanding of the function `evaluator::trim_objects` and support for evaluating Terraform modules' naming conventions
+- Comments for improving the understanding of the function `finder::trim_objects` and support for evaluating Terraform modules' naming conventions
 ### Changed
 Nothing
 ### Fixed